BOA: Do not try to add o_contrib_seven in D8 (Robert)

+++ aegir/tools/system/usage.sh
@@ -129,7 +129,7 @@ EOF
 fix_o_contrib_symlink()
 {
 if [ "$_O_CONTRIB" != "NO" ] ; then
-  if [ -e "$Plr/web.config" ] ; then
+  if [ -e "$Plr/web.config" ] && [ ! -e "$Plr/core" ] ; then
     if [ ! -e "$Plr/modules/o_contrib_seven" ] ; then
       ln -s $_O_CONTRIB_SEVEN $Plr/modules/o_contrib_seven
     fi
